Extract config to separate file
It’s gonna get larger eventually so might as well do it now
