+++ packages/tailwindcss/src/plugin-api.ts
@@ -1,6 +1,7 @@
 import { substituteAtApply } from './apply'
 import { decl, rule, type AstNode } from './ast'
 import type { NamedUtilityValue } from './candidate'
+import { createCompatConfig } from './compat/config/create-compat-config'
 import { deepMerge } from './compat/config/deep-merge'
 import { resolveConfig } from './compat/config/resolve-config'
 import type { UserConfig } from './compat/config/types'
@@ -9,7 +10,6 @@ import type { Theme } from './theme'
 import { withAlpha, withNegative } from './utilities'
 import { DefaultMap } from './utils/default-map'
 import { inferDataType } from './utils/infer-data-type'
-import { segment } from './utils/segment'
 import { toKeyPath } from './utils/to-key-path'
 
 export type Config = UserConfig
@@ -270,7 +270,7 @@ export function registerPlugins(plugins: Plugin[], designSystem: DesignSystem, a
 
   // Now merge all the configs and make all that crap work
   let resolvedConfig = resolveConfig([
-    createCompatabilityConfig(designSystem.theme),
+    createCompatConfig(designSystem.theme),
     ...pluginObjects.map(({ config }) => config ?? {}),
   ])
 
@@ -412,180 +412,6 @@ function set(obj: any, path: string[], value: any) {
 
 let themeUpgradeMap: Record<string, string[]> = {
   colors: ['color'],
-}
-
-function bareValues(fn: (value: NamedUtilityValue) => string | undefined) {
-  return {
-    // Ideally this would be a Symbol but some of the ecosystem assumes object with
-    // string / number keys for example by using `Object.entries()` which means that
-    // the function that handles the bare value would be lost
-    __BARE_VALUE__: fn,
-  }
-}
-
-let bareIntegers = bareValues((value) => {
-  if (!Number.isNaN(Number(value.value))) {
-    return value.value
-  }
-})
-
-let barePercentages = bareValues((value: NamedUtilityValue) => {
-  if (!Number.isNaN(Number(value.value))) {
-    return `${value.value}%`
-  }
-})
-
-let barePixels = bareValues((value: NamedUtilityValue) => {
-  if (!Number.isNaN(Number(value.value))) {
-    return `${value.value}px`
-  }
-})
-
-let bareMilliseconds = bareValues((value: NamedUtilityValue) => {
-  if (!Number.isNaN(Number(value.value))) {
-    return `${value.value}ms`
-  }
-})
-
-let bareDegrees = bareValues((value: NamedUtilityValue) => {
-  if (!Number.isNaN(Number(value.value))) {
-    return `${value.value}deg`
-  }
-})
-
-function createCompatabilityConfig(theme: Theme): UserConfig {
-  return {
-    theme: {
-      accentColor: ({ theme }) => theme('colors'),
-      aspectRatio: bareValues((value) => {
-        if (value.fraction === null) return
-        let [lhs, rhs] = segment(value.fraction, '/')
-        if (!Number.isInteger(Number(lhs)) || !Number.isInteger(Number(rhs))) return
-        return value.fraction
-      }),
-      backdropBlur: ({ theme }) => theme('blur'),
-      backdropBrightness: ({ theme }) => ({
-        ...theme('brightness'),
-        ...barePercentages,
-      }),
-      backdropContrast: ({ theme }) => ({
-        ...theme('contrast'),
-        ...barePercentages,
-      }),
-      backdropGrayscale: ({ theme }) => ({
-        ...theme('grayscale'),
-        ...barePercentages,
-      }),
-      backdropHueRotate: ({ theme }) => ({
-        ...theme('hueRotate'),
-        ...bareDegrees,
-      }),
-      backdropInvert: ({ theme }) => ({
-        ...theme('invert'),
-        ...barePercentages,
-      }),
-      backdropOpacity: ({ theme }) => ({
-        ...theme('opacity'),
-        ...barePercentages,
-      }),
-      backdropSaturate: ({ theme }) => ({
-        ...theme('saturate'),
-        ...barePercentages,
-      }),
-      backdropSepia: ({ theme }) => ({
-        ...theme('sepia'),
-        ...barePercentages,
-      }),
-      backgroundColor: ({ theme }) => theme('colors'),
-      backgroundOpacity: ({ theme }) => theme('opacity'),
-      border: barePixels,
-      borderColor: ({ theme }) => theme('colors'),
-      borderOpacity: ({ theme }) => theme('opacity'),
-      borderSpacing: ({ theme }) => theme('spacing'),
-      boxShadowColor: ({ theme }) => theme('colors'),
-      brightness: barePercentages,
-      caretColor: ({ theme }) => theme('colors'),
-      columns: bareIntegers,
-      contrast: barePercentages,
-      divideColor: ({ theme }) => theme('borderColor'),
-      divideOpacity: ({ theme }) => theme('borderOpacity'),
-      divideWidth: ({ theme }) => ({
-        ...theme('borderWidth'),
-        ...barePixels,
-      }),
-      fill: ({ theme }) => theme('colors'),
-      flexBasis: ({ theme }) => theme('spacing'),
-      flexGrow: bareIntegers,
-      flexShrink: bareIntegers,
-      gap: ({ theme }) => theme('spacing'),
-      gradientColorStopPositions: barePercentages,
-      gradientColorStops: ({ theme }) => theme('colors'),
-      grayscale: barePercentages,
-      gridRowEnd: bareIntegers,
-      gridRowStart: bareIntegers,
-      gridTemplateColumns: bareValues((value) => {
-        if (!Number.isNaN(Number(value.value))) {
-          return `repeat(${value.value}, minmax(0, 1fr))`
-        }
-      }),
-      gridTemplateRows: bareValues((value) => {
-        if (!Number.isNaN(Number(value.value))) {
-          return `repeat(${value.value}, minmax(0, 1fr))`
-        }
-      }),
-      height: ({ theme }) => theme('spacing'),
-      hueRotate: bareDegrees,
-      inset: ({ theme }) => theme('spacing'),
-      invert: barePercentages,
-      lineClamp: bareIntegers,
-      margin: ({ theme }) => theme('spacing'),
-      maxHeight: ({ theme }) => theme('spacing'),
-      maxWidth: ({ theme }) => theme('spacing'),
-      minHeight: ({ theme }) => theme('spacing'),
-      minWidth: ({ theme }) => theme('spacing'),
-      opacity: barePercentages,
-      order: bareIntegers,
-      outlineColor: ({ theme }) => theme('colors'),
-      outlineOffset: barePixels,
-      outlineWidth: barePixels,
-      padding: ({ theme }) => theme('spacing'),
-      placeholderColor: ({ theme }) => theme('colors'),
-      placeholderOpacity: ({ theme }) => theme('opacity'),
-      ringColor: ({ theme }) => theme('colors'),
-      ringOffsetColor: ({ theme }) => theme('colors'),
-      ringOffsetWidth: barePixels,
-      ringOpacity: ({ theme }) => theme('opacity'),
-      ringWidth: barePixels,
-      rotate: bareDegrees,
-      saturate: barePercentages,
-      scale: barePercentages,
-      scrollMargin: ({ theme }) => theme('spacing'),
-      scrollPadding: ({ theme }) => theme('spacing'),
-      sepia: barePercentages,
-      size: ({ theme }) => theme('spacing'),
-      skew: bareDegrees,
-      space: ({ theme }) => theme('spacing'),
-      stroke: ({ theme }) => theme('colors'),
-      strokeWidth: barePixels,
-      textColor: ({ theme }) => theme('colors'),
-      textDecorationColor: ({ theme }) => theme('colors'),
-      textDecorationThickness: barePixels,
-      textIndent: ({ theme }) => theme('spacing'),
-      textOpacity: ({ theme }) => theme('opacity'),
-      textUnderlineOffset: barePixels,
-      transitionDelay: bareMilliseconds,
-      transitionDuration: {
-        DEFAULT: theme.get(['--default-transition-duration']) ?? null,
-        ...bareMilliseconds,
-      },
-      transitionTimingFunction: {
-        DEFAULT: theme.get(['--default-transition-timing-function']) ?? null,
-      },
-      translate: ({ theme }) => theme('spacing'),
-      width: ({ theme }) => theme('spacing'),
-      zIndex: bareIntegers,
-    },
-  }
 }
 
 export type CssInJs = { [key: string]: string | CssInJs }
